Rg 25 (#21)
Implement open/close variables section and add responsive design


![image](https://github.com/ImmelstronDev/GraphQL-app/assets/106100393/c1a83192-bc5c-411c-9ebf-20173a4c9f2e)


![image](https://github.com/ImmelstronDev/GraphQL-app/assets/106100393/35f3685c-5f9e-41c5-81a8-8fb11ff33cd3)
