+++ public/locales/en/common.json
@@ -24,13 +24,16 @@
   "Logout": "Logout",
   "variables": "Variables",
   "headers": "Headers",
-  "error-boundary": "Sorry.. there was an error, try one more time",
+  "errorBoundary": "Sorry... there was an error, try one more time",
   "year": "© 2023",
+  "parseingError": "You need to get the API schema make a valid request to get the 'data'",
   "ForgotPassword": "Forgot Password",
   "DoNotHaveAnAccount": " Do not have an account?",
   "Register": "Register",
   "now": "now.",
   "Registered": "Register",
   "RegisterWithGoogle": "Register with Google",
-  "AlreadyHaveAnAccount": "Already have an account?"
+  "AlreadyHaveAnAccount": "Already have an account?",
+  "editorPlaceholder": "Write GraphQl request here",
+  "variablesPlaceholder": "Write GraphQl Variables here"
 }
