+++ public/locales/ru/common.json
@@ -24,13 +24,16 @@
   "Logout": "Выйти",
   "variables": "Переменные",
   "headers": "Заголовки",
-  "error-boundary": "Извините... произошла ошибка, попробуйте ещё раз",
+  "errorBoundary": "Извините... произошла ошибка, попробуйте ещё раз",
   "year": "© 2023",
+  "parseingError": "Вы должны получить схему API и сделать корректный запрос, для получения данных",
   "ForgotPassword": "Забыли пароль",
   "DoNotHaveAnAccount": " Не зарегистрированны?",
   "Register": "Зарегистируйтесь",
   "now": "сейчас.",
   "Registered": "Зарегистрироваться",
   "RegisterWithGoogle": "Зарегистрироваться чреез Google",
-  "AlreadyHaveAnAccount": "Уже есть аккаунт?"
+  "AlreadyHaveAnAccount": "Уже есть аккаунт?",
+  "editorPlaceholder": "Напишите GraphQL запрос здесь",
+  "variablesPlaceholder": "Напишите GraphQL переменные здесь"
 }
