+++ src/widgets/layouts/main/ui/Footer/Footer.module.css
@@ -5,6 +5,12 @@
   justify-content: space-around;
 
   width: 100%;
+  padding: 5px;
+
+  @media (width <= 750px) {
+    flex-direction: column;
+    row-gap: 15px;
+  }
 }
 
 .git {
