+++ src/widgets/layouts/main/ui/Header/Header.tsx
@@ -1,4 +1,5 @@
 import { Box } from '@mui/material';
+import { useRouter } from 'next/router';
 import { FC, useEffect, useState } from 'react';
 
 import { LangSwitcher } from '@/shared/ui';
@@ -11,6 +12,7 @@ type HeaderType = {
 
 export const Header: FC<HeaderType> = () => {
   const [scrolled, setScrolled] = useState(false);
+  const { pathname } = useRouter();
 
   useEffect(() => {
     const handleScroll = () => {
@@ -30,7 +32,7 @@ export const Header: FC<HeaderType> = () => {
   return (
     <Box component="header" className={className}>
       <div className={styles.wrapper}>
-        <LangSwitcher />
+        {pathname === '/' && <LangSwitcher />}
         <AuthContainer />
       </div>
     </Box>
