+++ src/widgets/layouts/main/ui/MainLayout/layout.module.css
@@ -26,6 +26,6 @@
 
   width: 100%;
   height: 100%;
-  margin-top: 100px;
+  margin-top: 60px;
   padding: 5px;
 }
