+++ src/widgets/layouts/side-bar/ui/Sidebar/Sidebar.module.css
@@ -1,6 +1,13 @@
 .container {
+  display: flex;
+  flex-direction: column;
   padding-right: 10px;
   padding-left: 10px;
+
+  @media (width <= 750px) {
+    flex-direction: row;
+    justify-content: center;
+  }
 }
 
 .icon {
